pyproject.toml
[project]
name = "django-starter"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2,<5.3",
    "gunicorn>=22.0",
    "psycopg[binary,pool]>=3.1",
    "whitenoise[brotli]>=6.7",
    "django-environ>=0.11",
    "argon2-cffi>=23.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=4.4",
    "ruff>=0.6",
    "pytest>=8.0",
    "pytest-django>=4.8",
    "coverage>=7.6",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["tests.py", "test_*.py"]
